Remove legacy svg.ts and cleanup codebase
Removed the old SVG renderer after moving everything to the new modular renderer
