+++ src/lib/svg/index.ts
@@ -1,0 +1,4 @@
+export { renderCard } from "./stats-card";
+export { renderErrorCard } from "./svg-primitives";
+export { renderLanguageChart } from "./languages";
+export { renderSparkline, type SparklineOptions } from "./sparkline";
