+++ src/lib/svg/languages/types/types.ts
@@ -1,0 +1,13 @@
+import {
+  ThemeConfig,
+  LanguageStat,
+  LangChartLayout,
+  LangChartOptions,
+} from "@/lib/types";
+
+export type LanguageChartRenderer = (
+  languages: LanguageStat[],
+  totalSize: number,
+  theme: ThemeConfig,
+  options: LangChartOptions,
+) => string;
