+++ src/lib/svg/text-metrics.ts
@@ -1,0 +1,48 @@
+export function charWidth(ch: string, fontSize: number): number {
+  const c = ch.toLowerCase();
+  if ("ijlt!:;|".includes(c)) return fontSize * 0.27;
+  if ("frskce".includes(c)) return fontSize * 0.42;
+  if ("adgnoquvxyzsbhkp".includes(c)) return fontSize * 0.5;
+  if ("mwW".includes(c)) return fontSize * 0.58;
+  if ("MQ@".includes(c)) return fontSize * 0.69;
+  return fontSize * 0.5;
+}
+
+export function estimateTextWidth(text: string, fontSize: number): number {
+  let width = 0;
+  for (const ch of text) width += charWidth(ch, fontSize);
+  return width;
+}
+
+export function truncateToWidth(
+  text: string,
+  maxWidth: number,
+  fontSize: number,
+): string {
+  let width = 0;
+  let lastFit = 0;
+  for (let i = 0; i < text.length; i++) {
+    const w = charWidth(text[i], fontSize);
+    if (width + w > maxWidth) break;
+    width += w;
+    lastFit = i + 1;
+  }
+  if (lastFit === text.length) return text;
+  if (lastFit === 0) return "\u2026";
+  const ellipsisW = charWidth("\u2026", fontSize);
+  while (lastFit > 0 && width + ellipsisW > maxWidth) {
+    width -= charWidth(text[lastFit - 1], fontSize);
+    lastFit--;
+  }
+  if (lastFit === 0) return "\u2026";
+  return text.slice(0, lastFit) + "\u2026";
+}
+
+/** Formats a language's share of the total as a percentage string, with a
+ * floor so vanishingly small (but nonzero) shares don't display as "0.0%". */
+export function formatLangPct(size: number, total: number): string {
+  if (total <= 0) return "0.0%";
+  const pct = (size / total) * 100;
+  if (pct > 0 && pct < 0.05) return "0.1%";
+  return `${pct.toFixed(1)}%`;
+}
